rearrange contracts folder
