--- contracts/abi/dsProxy/addresses.json
@@ -1,6 +1,0 @@
-{
-  "DS_PROXY_FACTORY": "0x23f67a19dc232835eaeda2075728f8295f54dfca",
-  "PROXY_REGISTRY": "0x9706786bf567796647d9428076fbc219830116ae",
-  "DS_PROXY": "0xaff08328e5a586754702f570d70972c43ab82ef8",
-  "SAI_PROXY": "0xc72b03c37735cf122c27dc352e5f25f75beea389"
-}
