+++ contracts/addresses/testnet.json
@@ -1,22 +1,22 @@
 {
-  "GEM": "0x7ba25f791fa76c3ef40ac98ed42634a8bc24c238",
-  "GOV": "0x1c3ac7216250edc5b9daa5598da0579688b9dbd5",
-  "PIP": "0xb7092ee7a8c4c85431962662310bbdcd4fd519e9",
-  "PEP": "0xc0ee05307ae4a5316f34874a3525d10c94b3c217",
+  "GEM": "0x70948167bf1ce7e80092bd5f0f1920720f26b705",
+  "GOV": "0x541e280ba5a2003bece51a1d7440e29f31a73bef",
+  "PIP": "0xe98c1b66e71eaa3dc942f120b0318cb38907d660",
+  "PEP": "0x958d30564db7cf3d2df452e1f110ad6abad6b610",
   "PIT": "0x0000000000000000000000000000000000000123",
-  "ADM": "0x4986c24c7f752c2ac2d738f1270639dd9e9d7bf5",
-  "SAI": "0xc226f3cd13d508bc319f4f4290172748199d6612",
-  "SIN": "0xe9e2b40d676fc998ede8c676d9f529ccbbc13740",
-  "SKR": "0xa6164a2e88e258a663772ed4912a0865af8f6d06",
-  "DAD": "0x7b61731911e46da837e3dcd2d8797de684c8ced1",
-  "MOM": "0x603d52d6ae2b98a49f8f32817ad4effe7e8a2502",
-  "VOX": "0xe16bf7aafeb33cc921d6d311e0ff33c4faa836dd",
-  "TUB": "0xe82ce3d6bf40f2f9414c8d01a35e3d9eb16a1761",
-  "TAP": "0x6896659267c3c9fd055d764327199a98e571e00d",
-  "TOP": "0x2774031b3898fbe414f929b3223ce1039325e7dc",
-  "MAKER_OTC": "0x06ef37a95603cb52e2dff4c2b177c84cdb3ce989",
-  "DS_PROXY_FACTORY": "0x23f67a19dc232835eaeda2075728f8295f54dfca",
-  "PROXY_REGISTRY": "0x9706786bf567796647d9428076fbc219830116ae",
-  "DS_PROXY": "0xaff08328e5a586754702f570d70972c43ab82ef8",
-  "SAI_PROXY": "0xc72b03c37735cf122c27dc352e5f25f75beea389"
+  "ADM": "0xddf2ecbb08893ae16a9ec2aba5f05a7c9d97cd88",
+  "SAI": "0x476f347cbffd82d21848957e4bb89e6179b0ee39",
+  "SIN": "0xb3aea1d9b5ef158fa87becc9f2795ebda6305301",
+  "SKR": "0xf84b804fdb917b74c862d493a02ffcaf132f6b38",
+  "DAD": "0x9edf18b7b562a17d9f458bff758c4544ea26782b",
+  "MOM": "0xda04731f1df8e15e65a40ed581fb79da701e0054",
+  "VOX": "0x83762de60969769599d55b1cec096fbccec5f65a",
+  "TUB": "0xdb3a2a7919b268eb556619566da1313abcbb265a",
+  "TAP": "0xf2ee3725840ec8282a25d94eb578bc45c41a046c",
+  "TOP": "0xcc1911c6b9c53bbd81f2293e28dc23536fddc129",
+  "MAKER_OTC": "0x07759b3b4201b224c27448e59ef6bf3c3d89a098",
+  "DS_PROXY_FACTORY": "0x135856a961dd102c11336d07c2ec00d2d0d4d077",
+  "PROXY_REGISTRY": "0xdcfb03e3aacc25c8cecd7d5ce17f09b61f232883",
+  "DS_PROXY": "0xbc496eb457ad94fbd7f55b1ef956583f1a5e3b57",
+  "SAI_PROXY": "0x33315f090e6a850fb7f482462251c23bdf3fbe78"
 }
